Rename calculate-aqi.py to AQICalculator.py
