+++ common/__init__.py
@@ -1,2 +1,2 @@
-from common.calculate_aqi import AQICalculator
+from common.AQICalculator import AQICalculator
 from common.connect_remote_db import connect_remote_db
